Added a new screen lock command that also powers down the monitor
I have chronic problems with dpms power management, the open-source
Intel graphics drivers and several monitor models I've used. The
behavior manifests in the monitor not powering down when xscreensaver
commands it to, sometimes staying on overnight. This is bad for the
monitor backlighting and therefore unacceptable.

A disappointing workaround has been to manually start the screensaver
(to lock the workstation) and then immediately executing `xset dpms
force off`. I'm now doing this so often I baked it into a new xmonad key
command, which is what this commit does.